Revert "Split GridFS bucket tests to improve performance of unit tests"
This reverts commit 1d831e2ebc74e373fe5321a02ed0005c944a75c7.
